requirements.txt
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
fastapi==0.110.2
h11==0.14.0
httptools==0.6.1
idna==3.7
motor==3.4.0
numpy==1.26.4
pandas==2.2.2
pydantic==2.7.1
pydantic_core==2.18.2
pymongo==4.7.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
tinydb==4.8.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.29.0
watchfiles==0.21.0
websockets==12.0
